Fetching Logic

+++ app/src/main/java/com/example/grocerywise/categorizationResponse.kt
@@ -10,7 +10,7 @@ data class CtgResponse(
 )
 
 data class Ingredient(
-    @Json(name = "amount") val amount: Int,
+    @Json(name = "amount") val amount: Double,
     @Json(name = "unit") val unit: String?,
     @Json(name = "name") val name: String,
 )
